% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_dataframe_p2.R
\name{generate_dataframe}
\alias{generate_dataframe}
\title{Generate dataframe ia a package that will help to create predecessors
files along with empty assigned variables arranged in an order.}
\usage{
generate_dataframe(required_files, previous_function_variable)
}
\description{
A csv file containing information about variables from metadataviewer
file is required.
First, load the csv file containing only variables from the metadaviewer file.
The first step is using: metadata() function.
syntax: metadata(metadata_file, create)
metadata_file: The csv file containing the information about varibles form
metadataviewer file.
create       : The dataset you want to create(Example: 'adae')
Using function metadata(), you will be asked to import certain files from sdtm or adam.
Then, another function is used to create the dataset, The newly imported dataset from adam or sdtm should be
inserted in the function in a sequence stated by the program as a list.
}
\details{
The function to be used is: generate_dataframe(required_files = list(), previous_function_variable)
required_files            : After using the previous function, the program will ask to insert the files in the given sequence
and it should be inserted as a list: example: generate_dataframe(required_files = list(a, b, c, ..), )
previous_function_variable: It is the variable that you used for storing the metadata() function.
}
